feat: new src file for dropsonde wind analysis
